service_embedding/requirements.txt
sentence-transformers
python-dotenv
flask